requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
pandas
openpyxl